Update and rename lib/lib.cpp to либ/lib.cpp
